Dwarf5 line number programs. (#775)
* Provide improved DWARF5 line number program parsing.

* Add new line number header descriptor class.

* Update SARIF SDK to improve file load exception reporting.

* Whitespace changes.

* Further updates while investigating apparent issues parsing MIPS assembler-generated scan targets.

* Update line number parsing behaviors to break under catastrophic conditions.

* Formatting changes.

* Complete file information parsing.

* Only parse maximum operations per instruction if available.

+++ src/BinaryParsers/ElfBinary/Dwarf/DwarfLineInformation.cs
@@ -26,6 +26,6 @@ namespace Microsoft.CodeAnalysis.BinaryParsers.Dwarf
         /// <summary>
         /// Gets or sets the column.
         /// </summary>
-        public uint Column { get; set; }
+        public ulong Column { get; set; }
     }
 }
